Add relevancy experiments
Support 3 relevancy variants (top, engaging, related) on the comment
threads for r/gaming. When a non-control variant is active, show only 3
comments (with an option to see the rest), and show 3 items of relevant
content below the comments.

Support "and" and explicit "or" feature config constructs, so that we
can condition on combinations of experiment variants, url flags,
subreddits, etc.

Support experiment variants in the feature flag component.

Add a RelevantContent component to encapsulate the content for the 3
variants.

We send bucketing events when determining which variant we're in, and we
send click events for clicks in the experiment area.