Allow root domain to be localhost
Problem:
In dev mode, setting cookies was not working because they're dependent on
your domain. Because we're setting cookies based off a hardcoded value in
development, testing cookies was difficult, since your local dev
environment's domain is different than the hardcoded one.

Fix:
Allow for the environment variable REDDIT to be something like
`localhost` or `foobar.local`. This works by checking for the last two
elements in a domain.  If those elements exist, use those for the root
domain. Ignore everything else but use those pieces to ensure it's a
fully qualified domain (scheme and a domain). If it doesn't find those
pieces, it just falls back to whatever you set as REDDIT. I also think
the regex here simplifies the code a bit.