Fix a bug in screenview tracking
Problem:
When shell was set to false, we were server rendering. Also, since shell
was set to false, we were running a function that was building tracking
data for us. This function relied on the window object, which obviously
doesn't exist server-side, and was causing an exception.

Fix:
Because we want to make a client side check before we build this data,
I'm now passing the data building function as a callback to the
`logClientScreenView` function.