Comment collapse/expand animation
This leverages react-motion to create a nice collapse/expand animation
when you tap the comment header. This animates both the height and the
opacity of the comment tree.

Details:
This works by wrapping everything that we want to animate in a div which
I called `Comment__base` in the css - this is the comment content, the
footer, and the replies. Once I've done this I wrap it in yet another
div. This div will effectively act as a mask whose overflow is hidden,
interpolating from the base's full height to 0 on collapse and 0 to the
base's full height on the expand - making it look as if the comment
itself is expanding or collapsing.

By doing it this way, the height of the comment base is always available
to be measured. For instance, if a comment starts off in a collapsed
state, to get its real height, you'd have to unhide/measure/rehide. This
scheme avoids a lot of that awkward handling.

The need for measuring the height in actual pixels is due to the
unfortunate fact that you can't interpolate from 0 to auto, a div's
default layout height, so a measurement must be taken before animation
begins. This is done simply in `componentWillUpdate`.

As a final detail, once the comment is finished animating, I switch the
height back to `auto`. This is a passive and simple way to deal with any
child elements changing their own heights.