Add reporting modal to comments/posts
Feature:
A modal that pops up when a user tries to report a comment or a post
(messages will come in a followup patch). Upon successful reporting, the
modal will close and a toaster pops up to notify the user all went well.
Clicking anywhere outside the modal box will also close modal.

Details:
This feature works by having a top level modal component. Inside this
component, there is a switch statement checking for modal types in the
`modal` redux state key. The modal key also carries with it the props
intended for the modal. In this case, just the `thingId` is necessary to
report someone so that's the only prop passed into `ReportingModal`.

To trigger the modal, the modal reducer is listening for the REPORT
action, which will set the report modal type.

Other notes:
- I've moved `DropdownCover` into `AppMain.jsx`. Since this is a UI component, this feels like a more appropriate place for it.

Here's what the modal looks like:
![reporting modal](https://cloud.githubusercontent.com/assets/787203/19126773/2651db52-8af1-11e6-8d43-0eb2d10c076e.gif)