Replying overhaul
This is an overhaul to replying incorporating our best practices. Which
means:

1. Real ajax requests using JSForm / removal of the reply POST handler.
2. A success and failure toaster.
3. Simplified reducer code for handling reply state.

Details:
This started off with a simple removal of the redirect that was
happening in the POST handler but I took it as an opportunity to clean
some of the older code (when we were still forming ideas on how this app
should work.)

There were two reducers addressing reply state. One seemed to handle the
view and one seemed to handle fetching. Nothing seemed to be taking
advantage of the latter so I just nixed it entirely. You can see
replying works fine without it. The `replying` reducer now only handles
whether or not a user is currently replying to a comment.

As mentioned above, I've also added a success toaster which triggers if
a reply was successful. The pre-existing failure toaster triggers if
not. One thing to note is that this is not optimistic. This is due to
needing the canonical reply model from the api to update the state. I
think this is fine as from a user's perspective, it's not completely
unexpected behavior.

Caveats:
- This is dependent on this [node-api-client
  patch](https://github.com/reddit/node-api-client/pull/171). Without
  it, a user is unable to reply to a reply unless he or she refreshes.

What this looks like:
![reply overhaul](https://cloud.githubusercontent.com/assets/787203/19630923/f9b64086-9948-11e6-8c8e-483d2c0c8b22.gif)