Posts and comments - different date formats
For 7 days we show the "time ago" format with increasing increments,
minutes to hours to days, then after that we show the actual date.

How this looks:
- "r/foobarsub ● 5 mins ● u/thephilthe"
- "r/foobarsub ● 8 hrs ● u/thephilthe"
- "r/foobarsub ● 1 day ● u/thephilthe"
- "r/foobarsub ● 7 days ● u/thephilthe"
- "r/foobarsub ● Aug 18 ● u/thephilthe"
- "r/foobarsub ● Dec 20, 2015 ● u/thephilthe"

On posts on listings, this can get a little cramped (it's already pretty
cramped), so I took a page from the iOS app and removed the user. The
user will show up on the comments page version of the Post.